Move Button Content to new component
